Revert "Fix some whitespace issue and other code reformatting"
This reverts commit a119a34cf74b9eabf5311aee2946b6eadebd5a25.

+++ app/src/main/java/com/demasu/testpixeldungeon/items/wands/WandOfLightning.java
@@ -69,7 +69,7 @@ public class WandOfLightning extends Wand {
         }
 
         affected.add( ch );
-        ch.damage( Level.water[ch.pos] && !ch.flying ? damage * 2 : damage, LightningTrap.LIGHTNING );
+        ch.damage( Level.water[ch.pos] && !ch.flying ? (int) ( damage * 2 ) : damage, LightningTrap.LIGHTNING );
 
         ch.sprite.centerEmitter().burst( SparkParticle.FACTORY, 3 );
         ch.sprite.flash();
